requirements.txt
fastapi==0.75.2
numpy==1.22.2
pandas==1.4.1
pymongo==4.0.1
pydantic==1.9.0
PyYAML==6.0
sentry-sdk==1.5.10
python-dotenv==0.19.2
requests~=2.27.1
uvicorn==0.15.0
python-dateutil~=2.8.2
redis~=4.3.4
cronitor~=4.5.0
apscheduler~=3.9.1
scikit-learn~=1.0.2
tqdm~=4.64.1
matplotlib~=3.6.2
mlxtend~=0.20.0
locust~=2.11.1
psycopg2~=2.9.3
